third_party/blink/web_tests/http/tests/eventsource/eventsource-reconnect-on-redirect-expected.txt

Test EventSource reconnect after end of event stream.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS es.readyState is es.CONNECTING
PASS es.readyState is es.OPEN
got a message: url = /eventsource/resources/echo-url.php?id=, id = 
PASS es.readyState is es.CONNECTING
PASS es.readyState is es.OPEN
got a message: url = /eventsource/resources/echo-url.php?id=, id = 77
PASS es.readyState is es.CLOSED
PASS es.url is "http://127.0.0.1:8000/eventsource/resources/redirect.php"
PASS successfullyParsed is true

TEST COMPLETE
